+++ keyboard.h
@@ -1,7 +1,9 @@
 #ifndef KEYBOARD_H
 #define KEYBOARD_H
 
-void setup_keyboard(int);
+#include <stdint.h>  // For uint8_t.
+
+void setup_keyboard(uint8_t);
 void scan_keyboard();
 
 #endif
