+++ mcp3008.h
@@ -1,7 +1,9 @@
 #ifndef ANALOG_H
 #define ANALOG_H
 
-void mcp3008_setup(int);
-int mcp3008_read(int, int, int);
+#include <stdint.h>  // For uint8_t.
+
+void mcp3008_setup(uint8_t);
+unsigned int mcp3008_read(uint8_t, uint8_t, uint8_t);
 
 #endif
